high-requirements.txt
fastapi
uvicorn
redis
asyncpg
sqlmodel
yfinance
python-dotenv
aiosqlite
fakeredis
psycopg2-binary